avva says:
So, I did find one source of leak... but it's not the one that could
cause this bug, I think.

Look at stats of some memcache process that's been running long. In
almost all classes no. of items reported by 'stats items' is less than
"total chunks" of 'stats slabs' even though there aren't any free
slabs. Now the numbers aren't supposed to be equal, because some items
are out of linked list but tied up in delete queue or being printed.
But if the difference is in thousands of items, e.g. here:
http://www.livejournal.com/admin/temp/memcache.bml?host=10.0.0.32:11212&mode=raw
something is wrong.

Anyway, the bug I found is that when delete_handler() goes over the
queue, it decreases refcount rather than call item_remove(). In almost
all cases it won't make any difference, because the item is still in
the linked list and item_remove() wouldn't free it anyway. But if:

1) someone DELETEs an item and this sets its exptime to now()+4 and
puts it into the delete queue.
2) 4 seconds pass, but delete_handler() hasn't been called yet (it's
called every 5 seconds, approx.). In this brief and rare period of
time, this item is called with GET.
3) since its exptime passed, the item is unlinked. Now when
delete_handler() gets to it, it'll decrease the refcount, but noone
will ever free it.


He included a fix:

--- cvs/wcmtools/memcached/memcached.c  Tue Jun 24 10:38:55 2003
+++ src/memcached/memcached.c   Tue Jun 24 16:50:01 2003
@@ -1058,7 +1058,7 @@
             if (todelete[i]->exptime < now) {
                 /* no longer mark it deleted. it's now expired, same as dead */
                 todelete[i]->it_flags &= ~ITEM_DELETED;
-                todelete[i]->refcount--;
+                item_remove(todelete[i]);
             } else {
                 todelete[j++] = todelete[i];
             }


But I think what I'm commiting will work just as well.  (it just
removes items immediately instead of letting them fall off and die
later)


git-svn-id: http://code.sixapart.com/svn/memcached/trunk@42 b0b603af-a30f-0410-a34e-baf09ae79d0b