patch from David Phillips <electrum@gmail.com>
Here is a patch to configure.ac and Makefile.am to put the man page in
the correct location.  Trying to install the man page from a
subdirectory results in the subdirectory being used in the install
path (it tries to install to doc/memcached.1).  This is the correct
thing to do:

- create a Makefile.am in the doc directory that installs the man page
with man_MANS
- modify Makefile.am in the base directory to reference the doc
directory using SUBDIRS
- modify the AC_CONFIG_FILES macro in configure.ac to output the Makefile in doc


git-svn-id: http://code.sixapart.com/svn/memcached/trunk@234 b0b603af-a30f-0410-a34e-baf09ae79d0b